remove db_name from rds module (#76)
* remove db_name from rds module.

Removes the 'db_name' from our created rds database. This setting does create a default db when set. The problem is that the 'default db name' is immutable. When we want to do rename the database it won't let us to do without destroying and recreating it. The suggestion came from [here](https://stackoverflow.com/a/43106463) and we think the 'burden' of not being able to rename the database is larger/bigger/more impactful than having to create the database manually once.

---------

Co-authored-by: Merle Fischer <mer.fischer@gmail.com>