feature/automated redrive sqs (#163)
The sqs module now by default adds a lambda to the DLQ that automatically redrives messages from the DLQ to the created main queue. This is customizeable and by default runs once a day at 2AM UTC. If additional logic for the automated redrive lambda is required (e.g. filtering poisoned messages) the user has the option to disable automatically creating the redrive lambda.