fix sqs automated redrive lambda in yarn workspaces (#164)
Fixes build error when using sqs module with DLQ and automated redrive in yarn workspace by not building anymore and instead using the file from the dist folder via the is_bundle property. Error was along the lines of:


.terraform/modules/axxerion_sqs/aws/sqs/redrive-lambda)
│ doesn't seem to be part of the project declared in
│ /home/runner/work/<redacted>.
│ 
│ - If /home/runner/work/<redacted> isn't intended to be a
│ project, remove any yarn.lock and/or package.json file there.
│ - If /home/runner/work/<redacted> is intended to be a
│ project, it might be that you forgot to list
│ <redacted>/.terraform/modules/axxerion_sqs/aws/sqs/redrive-lambda
│ in its workspace configuration.
│ - Finally, if /home/runner/work/<redacted> is fine and
│ you intend
│ <redacted>/.terraform/modules/axxerion_sqs/aws/sqs/redrive-lambda
│ to be treated as a completely separate project (not even a workspace),
│ create an empty yarn.lock file in it.

This was tested by referencing the latest commit of this PR against the previously failing ci.